requirements.txt
numpy
tqdm
tiktoken
torch
datasets
huggingface-hub
hydra-core